+++ extension/json/src/functions/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_subdirectory(creation_functions)
+add_subdirectory(extract_functions)
+add_subdirectory(scalar_functions)
